pyproject.toml
[project]
name = "coderdojo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkslides>=1.0.10",
    "pygame-ce>=2.5.6",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
